Remove compiler warnings
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@788 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/dsutil/MediaTypeEx.cpp
@@ -156,7 +156,7 @@ CString CMediaTypeEx::ToString(IPin* pPin)
 			rtDur /= 60;
 			int m = rtDur%60;
 			rtDur /= 60;
-			int h = rtDur;
+			int h = (int)rtDur;
 			if(h) dur.Format(_T("%d:%02d:%02d"), h, m, s);
 			else if(m) dur.Format(_T("%02d:%02d"), m, s);
 			else if(s) dur.Format(_T("%ds"), s);
@@ -622,21 +622,21 @@ void CMediaTypeEx::Dump(CAtlList<CString>& sl)
 		{
 			str.Format(_T("%04x:"), i);
 
-			for(int k = i, l = min(i + 16, cbFormat); k < l; k++)
+			for(int k = i, l = min(i + 16, (int)cbFormat); k < l; k++)
 			{
 				CString byte;
 				byte.Format(_T("%c%02x"), fmtsize > 0 && fmtsize == k ? '|' : ' ', pbFormat[k]);
 				str += byte;
 			}
 
-			for(int k = min(i + 16, cbFormat), l = i + 16; k < l; k++)
+			for(int k = min(i + 16, (int)cbFormat), l = i + 16; k < l; k++)
 			{
 				str += _T("   ");
 			}
 
 			str += ' ';
 
-			for(int k = i, l = min(i + 16, cbFormat); k < l; k++)
+			for(int k = i, l = min(i + 16, (int)cbFormat); k < l; k++)
 			{
 				unsigned char c = (unsigned char)pbFormat[k];
 				CStringA ch;
